Avoid UNIV_PAGE_SIZE redefinition
